requirements.txt
streamlit
plotly
openai>=1.0.0
python-dotenv
